+++ packages/riverpod/lib/riverpod.dart
@@ -11,7 +11,15 @@ export 'src/async_notifier.dart'
         CancelAsyncSubscription,
         BuildlessAsyncNotifier,
         BuildlessAutoDisposeAsyncNotifier,
-        FutureHandlerProviderElementMixin;
+        FutureHandlerProviderElementMixin,
+        FamilyStreamNotifierProviderImpl,
+        StreamNotifierProviderImpl,
+        AutoDisposeStreamNotifierProviderImpl,
+        AutoDisposeFamilyStreamNotifierProviderImpl,
+        StreamNotifierProviderBase,
+        BuildlessAutoDisposeStreamNotifier,
+        BuildlessStreamNotifier;
+
 export 'src/common.dart' hide AsyncTransition;
 export 'src/framework.dart'
     hide