+++ packages/riverpod_generator/test/stream_notifier_test.dart
@@ -3,20 +3,22 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:test/test.dart';
 
-import 'integration/async.dart';
+import 'integration/stream.dart';
 import 'utils.dart';
 
 void main() {
   test(
-      'Creates an AsyncNotifierProvider<T> if @riverpod is used on an async class',
-      () {
+      'Creates a StreamNotifierProvider<T> if @riverpod is used on a Stream class',
+      () async {
     final container = createContainer();
 
-    final AutoDisposeAsyncNotifierProvider<PublicClass, String> provider =
+    final AutoDisposeStreamNotifierProvider<PublicClass, String> provider =
         publicClassProvider;
-    final AsyncValue<String> result = container.read(publicClassProvider);
 
-    expect(result, const AsyncData('Hello world'));
+    expect(
+      await container.listen(publicClassProvider.future, (_, __) {}).read(),
+      'Hello world',
+    );
   });
 
   test('Generates .name for providers', () {
@@ -29,7 +31,7 @@ void main() {
 
   test(
       'Creates a NotifierProvider.family<T> if @riverpod is used on a synchronous function with parameters',
-      () {
+      () async {
     final container = createContainer();
 
     const FamilyClassFamily family = familyClassProvider;
@@ -77,7 +79,7 @@ void main() {
       fifth: ['x42'],
     );
     // ignore: invalid_use_of_internal_member
-    final AutoDisposeAsyncNotifierProviderImpl<FamilyClass, String>
+    final AutoDisposeStreamNotifierProviderImpl<FamilyClass, String>
         futureProvider = provider;
 
     expect(provider.first, 42);
@@ -86,21 +88,20 @@ void main() {
     expect(provider.fourth, false);
     expect(provider.fifth, ['x42']);
 
-    final AsyncValue<String> result = container.read(
-      familyClassProvider(
-        42,
-        second: 'x42',
-        third: .42,
-        fourth: false,
-        fifth: ['x42'],
-      ),
-    );
-
     expect(
-      result,
-      const AsyncData(
-        '(first: 42, second: x42, third: 0.42, fourth: false, fifth: [x42])',
-      ),
+      await container
+          .listen(
+            familyClassProvider(
+              42,
+              second: 'x42',
+              third: .42,
+              fourth: false,
+              fifth: ['x42'],
+            ).future,
+            (_, __) {},
+          )
+          .read(),
+      '(first: 42, second: x42, third: 0.42, fourth: false, fifth: [x42])',
     );
   });
 }