+++ packages/riverpod/CHANGELOG.md
@@ -1,4 +1,20 @@
-## Unreleased minor
+## 2.3.0
+
+- Added `StreamNotifier` + `StreamNotifierProvider`.
+  This is for building a `StreamProvider` while exposing ways to modify the stream.
+
+  It is primarily meant to be used using code-generation via riverpod_generator,
+  by writing:
+
+  ```dart
+  @riverpod
+  class Example extends _$Example {
+    @override
+    Stream<Model> build() {
+      // TODO return some stream
+    }
+  }
+  ```
 
 - Deprecated `StreamProvider.stream`
   Instead of:
@@ -24,7 +40,7 @@
   Do:
 
   ```dart
-  final a = FuturProvider((ref) async {
+  final a = FutureProvider((ref) async {
     final e = await ref.watch(b.future);
     return Model(e);
   })
@@ -33,7 +49,8 @@
 - Some restrictions on the `dependencies` parameter of providers have been lifted.
   It is no-longer necessary to include providers which do not themselves specify `dependencies`.
   All providers should specify `dependencies` if they are scoped at any point.
-- Annotate `Notifier.state` setter as protected.
+
+- Annotated `Notifier.state` setter as protected.
 
 ## 2.2.0