+++ packages/riverpod_annotation/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+- Export necessary utilites for providers returning a Stream.
+- Upgraded riverpod dependency
+
 ## 1.2.1
 
 Bump minimum Riverpod version